fix broken tests and move cache package
