+++ interface/src/main/scala/quasar/main/Quasar.scala
@@ -26,9 +26,9 @@ import quasar.fp._
 import quasar.fp.free._
 import quasar.fp.numeric._
 import quasar.fs._
-import quasar.fs.cache.VCache
 import quasar.fs.mount._
 import quasar.fs.mount.BackendDef.DefinitionResult
+import quasar.fs.mount.cache.VCache
 import quasar.main.metastore._
 import quasar.metastore._
