+++ it/src/test/scala/quasar/fs/FileSystemTest.scala
@@ -22,8 +22,8 @@ import quasar.contrib.pathy._
 import quasar.contrib.scalaz.eitherT._
 import quasar.fp._
 import quasar.fp.free._
-import quasar.fs.cache.VCache
 import quasar.fs.mount._, BackendDef.DefinitionResult, Fixture._
+import quasar.fs.mount.cache.VCache
 import quasar.effect._
 import quasar.main.{KvsMounter, HierarchicalFsEffM, PhysFsEff, PhysFsEffM}
 import quasar.mimir
